requirements_tf2.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
dill==0.3.0
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
googleapis-common-protos==1.6.0
grpcio==1.22.0
h5py==2.9.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.16.4
promise==2.2.1
protobuf==3.8.0
psutil==5.6.3
pyparsing==2.4.0
python-dateutil==2.8.0
requests==2.22.0
scipy==1.3.0
six==1.12.0
tensorboard==1.14.0
tensorflow-datasets==1.0.2
tensorflow-estimator==1.14.0
tensorflow-gpu
tensorflow-metadata==0.13.0
termcolor==1.1.0
tqdm==4.32.2
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
